server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.2
faker==19.13.0
pandas==2.1.1
scikit-learn==1.3.1
numpy==1.26.0
matplotlib==3.8.0